python/json-formatter/requirements.txt
mcp>=1.0.0
pyyaml>=6.0
jsonschema>=4.20.0

